+++ Anarchy/REST/User/Settings/SettingsUpdatedEventArgs.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace Discord
+{
+    public class SettingsUpdatedEventArgs : EventArgs
+    {
+        public Settings Settings { get; private set; }
+
+        internal SettingsUpdatedEventArgs(Settings settings)
+        {
+            Settings = settings;
+        }
+    }
+}